Implement Anthropic models, mappers and provider

+++ backend/src/Routify.Gateway/Providers/TogetherAi/Models/TogetherAiCompletionInput.cs
@@ -36,7 +36,7 @@ internal record TogetherAiCompletionInput : ICompletionInput
     public float? RepetitionPenalty { get; set; }
     
     [JsonPropertyName("temperature")]
-    public double? Temperature { get; set; }
+    public float? Temperature { get; set; }
     
     [JsonPropertyName("logprobs")]
     public int? Logprobs { get; set; }
